LPS-145980 Move AuditGraphApp components to AuditGraphApp subfolder
