+++ modules/apps/content-dashboard/content-dashboard-web/src/main/resources/META-INF/resources/js/components/AuditGraphApp/AuditGraphApp.js
